+++ astropy/nddata/nddata.py
@@ -5,6 +5,7 @@ __all__ = ['NDData']
 
 import numpy as np
 
+from ..units import Unit
 from .flag_collection import FlagCollection
 from .nderror import IncompatibleErrorsException, NDError
 
@@ -25,7 +26,7 @@ class NDData(object):
 
     mask : `~numpy.ndarray`, optional
         Mask for the data, given as a boolean Numpy array with a shape
-        matching that of the data. The values should be ``False`` where the
+        matching that of the data. The values must be ``False`` where the
         data is *valid* and ``True`` when it is not (as for Numpy masked
         arrays).
 
@@ -49,13 +50,8 @@ class NDData(object):
         of this particular object.  e.g., creation date, unique identifier,
         simulation parameters, exposure time, telescope name, etc.
 
-    units : undefined, optional
+    units : `astropy.units.UnitBase` instance or str, optional
         The units of the data.
-
-        .. warning::
-            The units scheme is under development. For now, just supply a
-            string when relevant - the units system will likely be compatible
-            with providing strings to initialize itself.
 
     copy : bool, optional
         If True, the array will be *copied* from the provided `data`, otherwise
@@ -96,6 +92,8 @@ class NDData(object):
     def __init__(self, data, error=None, mask=None, flags=None, wcs=None,
                  meta=None, units=None, copy=True):
 
+        #TODO change constructor to also accept an nddata object (keywords overwrite current data)
+
         self.data = np.array(data, subok=True, copy=copy)
 
         self.error = error
@@ -119,14 +117,14 @@ class NDData(object):
         if value is not None:
             if isinstance(value, np.ndarray):
                 if value.dtype != np.bool_:
-                    raise TypeError("mask should be a boolean Numpy array")
+                    raise TypeError("mask must be a boolean Numpy array")
                 else:
                     if value.shape != self.shape:
                         raise ValueError("dimensions of mask do not match data")
                     else:
                         self._mask = value
             else:
-                raise TypeError("mask should be a Numpy array")
+                raise TypeError("mask must be a Numpy array")
         else:
             self._mask = value
 
@@ -163,9 +161,20 @@ class NDData(object):
                 self._error = value
                 self._error.parent_nddata = self
             else:
-                raise TypeError("error should be an instance of a NDError object")
+                raise TypeError("error must be an instance of a NDError object")
         else:
             self._error = value
+
+    @property
+    def units(self):
+        return self._units
+
+    @units.setter
+    def units(self, value):
+        if value is None:
+            self._units = None
+        else:
+            self._units = Unit(value)
 
     @property
     def shape(self):
@@ -205,19 +214,49 @@ class NDData(object):
         else:
             return self.data
 
-    # At the moment, the units are required to be the same for addition and
-    # subtraction. Once the units framework is implemented, we no longer have
-    # to have this requirement, and can also add methods for multiplication
-    # and division.
+    def __getitem__(self, item):
 
-    def add(self, operand, propagate_errors=True):
+        new_data = self.data[item]
+
+        if self.error is not None:
+            new_error = self.error[item]
+        else:
+            new_error = None
+
+        if self.mask is not None:
+            new_mask = self.mask[item]
+        else:
+            new_mask = None
+
+        if self.flags is not None:
+            if isinstance(self.flags, np.ndarray):
+                new_flags = self.flags[item]
+            elif isinstance(self.flags, FlagCollection):
+                raise NotImplementedError('Slicing complex Flags is currently not implemented')
+        else:
+            new_flags = None
+
+        if self.wcs is not None:
+            raise NotImplementedError('Slicing for WCS is not currently implemented')
+        else:
+            new_wcs = None
+
+        return self.__class__(new_data, error=new_error, mask=new_mask, flags=new_flags, wcs=new_wcs,
+            meta=self.meta, units=self.units, copy=False)
+
+
+
+
+
+
+    def _arithmetic(self, operand, propagate_errors, name, operation):
         """
-        Add another dataset (`operand`) to this dataset.
+        {name} another dataset (`operand`) to this dataset.
 
         Parameters
         ----------
         operand : `~astropy.nddata.NDData`
-            The second operand in the operation a + b
+            The second operand in the operation a {operator} b
         propagate_errors : bool
             Whether to propagate errors following the propagation rules
             defined by the class used for the `error` attribute.
@@ -229,26 +268,34 @@ class NDData(object):
 
         Notes
         -----
-        This method requires the datasets to have identical WCS properties,
-        identical units, and identical shapes. Flags and meta-data get set to
-        None in the resulting dataset. Errors are propagated, although this
-        feature is experimental, and errors are assumed to be uncorrelated.
-        Values masked in either dataset before the operation are masked in
-        the resulting dataset.
+        This method requires the datasets to have identical WCS
+        properties, equivalent units, and identical shapes. Flags and
+        meta-data get set to None in the resulting dataset. The unit
+        in the result is the same as the unit in `self`. Errors are
+        propagated, although this feature is experimental, and errors
+        are assumed to be uncorrelated.  Values masked in either
+        dataset before the operation are masked in the resulting
+        dataset.
         """
-
         if self.wcs != operand.wcs:
             raise ValueError("WCS properties do not match")
 
-        if self.units != operand.units:
-            raise ValueError("operand units do not match")
+        if not (self.units is None and operand.units is None):
+            if (self.units is None or operand.units is None
+                or not self.units.is_equivalent(operand.units)):
+                raise ValueError("operand units do not match")
 
         if self.shape != operand.shape:
             raise ValueError("operand shapes do not match")
 
-        result = self.__class__(self.data + operand.data)  # in case we are dealing with an inherited type
+        if self.units is not None:
+            operand_data = operand.units.to(self.units, operand.data)
+        else:
+            operand_data = operand.data
+        data = operation(self.data, operand_data)
+        result = self.__class__(data)  # in case we are dealing with an inherited type
 
-        if not propagate_errors:
+        if propagate_errors is None:
             result.error = None
         elif self.error is None and operand.error is None:
             result.error = None
@@ -258,9 +305,15 @@ class NDData(object):
             result.error = self.error
         else:  # both self and operand have errors
             try:
-                result.error = self.error.propagate_add(operand, result.data)
+                method = getattr(self.error, propagate_errors)
+                result.error = method(operand, result.data)
             except IncompatibleErrorsException:
-                raise IncompatibleErrorsException("Cannot propagate errors of type {0:s} with errors of type {1:s} for addition".format(self.error.__class__.__name__, operand.error.__class__.__name__))
+                raise IncompatibleErrorsException(
+                    "Cannot propagate errors of type {0:s} with errors of "
+                    "type {1:s} for {2:s}".format(
+                        self.error.__class__.__name__,
+                        operand.error.__class__.__name__,
+                        name))
 
         if self.mask is None and operand.mask is None:
             result.mask = None
@@ -278,70 +331,72 @@ class NDData(object):
 
         return result
 
+    def add(self, operand, propagate_errors=True):
+        if propagate_errors:
+            propagate_errors = "propagate_add"
+        else:
+            propagate_errors = None
+        return self._arithmetic(
+            operand, propagate_errors, "addition", np.add)
+    add.__doc__ = _arithmetic.__doc__.format(name="Add", operator="+")
+
     def subtract(self, operand, propagate_errors=True):
+        if propagate_errors:
+            propagate_errors = "propagate_subtract"
+        else:
+            propagate_errors = None
+        return self._arithmetic(
+            operand, propagate_errors, "subtraction", np.subtract)
+    subtract.__doc__ = _arithmetic.__doc__.format(name="Subtract", operator="-")
+
+    def multiply(self, operand, propagate_errors=True):
+        if propagate_errors:
+            propagate_errors = "propagate_multiply"
+        else:
+            propagate_errors = None
+        return self._arithmetic(
+            operand, propagate_errors, "multiplication", np.multiply)
+    multiply.__doc__ = _arithmetic.__doc__.format(name="Multiply", operator="*")
+
+    def divide(self, operand, propagate_errors=True):
+        if propagate_errors:
+            propagate_errors = "propagate_divide"
+        else:
+            propagate_errors = None
+        return self._arithmetic(
+            operand, propagate_errors, "division", np.divide)
+    divide.__doc__ = _arithmetic.__doc__.format(name="Divide", operator="/")
+
+    def convert_units_to(self, unit):
         """
-        Subtract another dataset (`operand`) from this dataset.
+        Returns a new `NDData` object whose values have been converted
+        to a new unit.
 
         Parameters
         ----------
-        operand : `~astropy.nddata.NDData`
-            The second operand in the operation a - b
-        propagate_errors : bool
-            Whether to propagate errors following the propagation rules
-            defined by the class used for the `error` attribute.
+        unit : `astropy.units.UnitBase` instance or str
+            The unit to convert to.
 
         Returns
         -------
         result : `~astropy.nddata.NDData`
             The resulting dataset
 
-        Notes
-        -----
-        This method requires the datasets to have identical WCS properties,
-        identical units, and identical shapes. Flags and meta-data get set to
-        None in the resulting dataset. Errors are propagated, although this
-        feature is experimental, and errors are assumed to be uncorrelated.
-        Values masked in either dataset before the operation are masked in
-        the resulting dataset.
+        Raises
+        ------
+        UnitsException
+            If units are inconsistent.
         """
+        if self.units is None:
+            raise ValueError("No units specified on source data")
+        data = self.units.to(unit, self.data)
+        result = self.__class__(data)  # in case we are dealing with an inherited type
 
-        if self.wcs != operand.wcs:
-            raise ValueError("WCS properties do not match")
-
-        if self.units != operand.units:
-            raise ValueError("operand units do not match")
-
-        if self.shape != operand.shape:
-            raise ValueError("operand shapes do not match")
-
-        result = self.__class__(self.data - operand.data)  # in case we are dealing with an inherited type
-
-        if not propagate_errors:
-            result.error = None
-        elif self.error is None and operand.error is None:
-            result.error = None
-        elif self.error is None:
-            result.error = operand.error
-        elif operand.error is None:
-            result.error = self.error
-        else:  # both self and operand have errors
-            try:
-                result.error = self.error.propagate_subtract(operand, result.data)
-            except IncompatibleErrorsException:
-                raise IncompatibleErrorsException("Cannot propagate errors of type {0:s} with errors of type {1:s} for subtractition".format(self.error.__class__.__name__, operand.error.__class__.__name__))
-
-        if self.mask is None and operand.mask is None:
-            result.mask = None
-        elif self.mask is None:
-            result.mask = operand.mask
-        elif operand.mask is None:
-            result.mask = self.mask
-        else:  # combine masks as for Numpy masked arrays
-            result.mask = self.mask & operand.mask
-
+        result.error = self.error
+        result.mask = self.mask
         result.flags = None
         result.wcs = self.wcs
-        result.meta = None
-        result.units = self.units
+        result.meta = self.meta
+        result.units = unit
 
         return result