+++ astropy/units/format/generic.py
@@ -5,33 +5,9 @@ Handles a "generic" string format for units
 from __future__ import absolute_import, division, print_function, unicode_literals
 
 from fractions import Fraction
-import functools
 
 from .base import Base
 from . import utils
-
-
-DEBUG = False
-
-
-def _trace(func):
-    """
-    A utility decorator to help debug the parser.
-    """
-    def run(self, s, loc, toks):
-        if DEBUG:
-            print(func.__name__, toks, end=' ')
-            try:
-                result = func(self, s, loc, toks)
-            except Exception as e:
-                print("Exception: ", e.message)
-                raise
-            print(result)
-        else:
-            return func(self, s, loc, toks)
-        return result
-
-    return functools.update_wrapper(run, func)
 
 
 class Generic(Base):
@@ -157,27 +133,27 @@ class Generic(Base):
         return main
 
     @classmethod
-    @_trace
+    @utils._trace
     def _parse_unsigned_integer(cls, s, loc, toks):
         return int(toks[0])
 
     @classmethod
-    @_trace
+    @utils._trace
     def _parse_signed_integer(cls, s, loc, toks):
         return int(toks[0])
 
     @classmethod
-    @_trace
+    @utils._trace
     def _parse_integer(cls, s, loc, toks):
         return int(toks[0])
 
     @classmethod
-    @_trace
+    @utils._trace
     def _parse_floating_point(cls, s, loc, toks):
         return float(toks[0])
 
     @classmethod
-    @_trace
+    @utils._trace
     def _parse_factor(cls, s, loc, toks):
         if len(toks) == 1:
             return toks[0]
@@ -187,12 +163,12 @@ class Generic(Base):
             return float(toks[0]) * toks[1] ** float(toks[2])
 
     @classmethod
-    @_trace
+    @utils._trace
     def _parse_frac(cls, s, loc, toks):
         return Fraction(toks[0], toks[1])
 
     @classmethod
-    @_trace
+    @utils._trace
     def _parse_unit(cls, s, loc, toks):
         from astropy.extern import pyparsing as p
 
@@ -202,7 +178,7 @@ class Generic(Base):
             s, loc, "{0!r} is not a recognized unit".format(toks[0]))
 
     @classmethod
-    @_trace
+    @utils._trace
     def _parse_product_of_units(cls, s, loc, toks):
         if len(toks) == 1:
             return toks[0]
@@ -210,12 +186,12 @@ class Generic(Base):
             return toks[0] * toks[1]
 
     @classmethod
-    @_trace
+    @utils._trace
     def _parse_division_product_of_units(cls, s, loc, toks):
         return (toks[0] * toks[1]) / toks[2]
 
     @classmethod
-    @_trace
+    @utils._trace
     def _parse_factor_product_of_units(cls, s, loc, toks):
         if toks[0] != 1.0:
             return toks[0] * toks[1]
@@ -223,7 +199,7 @@ class Generic(Base):
             return toks[1]
 
     @classmethod
-    @_trace
+    @utils._trace
     def _parse_unit_with_power(cls, s, loc, toks):
         if len(toks) == 1:
             return toks[0]
@@ -231,7 +207,7 @@ class Generic(Base):
             return toks[0] ** toks[1]
 
     @classmethod
-    @_trace
+    @utils._trace
     def _parse_function(cls, s, loc, toks):
         from astropy.extern import pyparsing as p
 
@@ -246,7 +222,7 @@ class Generic(Base):
     def parse(self, s):
         from astropy.extern import pyparsing as p
 
-        if DEBUG:
+        if utils.DEBUG:
             print("parse", s)
 
         if '_unit_namespace' not in Generic.__dict__:
@@ -267,9 +243,14 @@ class Generic(Base):
 
     def _format_unit_list(self, units):
         out = []
+        units.sort(key=lambda x: self._get_unit_name(x[0]).lower())
+
         for base, power in units:
             if power == 1:
                 out.append(self._get_unit_name(base))
+            elif isinstance(power, Fraction):
+                out.append('{0}({1})'.format(
+                    self._get_unit_name(base), power))
             else:
                 out.append('{0}{1}'.format(
                     self._get_unit_name(base), power))